requirements.txt
Django==2.2.7
djangorestframework
djangorestframework-jwt
PyJWT
pytz
